Fix the bug of Date of birth on Android
The day of birth for android have the issue that both you birth year, month and day can be negative number, and also the time set for driving age is for people who born after 2004, which from now is less than 16, but in 2020 they will be turn into 16 yet they still will show the message of their not old enough.

Now this bug is fix. The date won't be able to show negative number and the problem of driver's underage is solved.